EXAM-34: Rename folders and remove user from authUserSlice
