--- furniture-client/src/features/rootSaga.ts
@@ -1,8 +1,0 @@
-import { all } from 'redux-saga/effects';
-
-import { userSaga } from './slice/user/userSaga';
-
-
-export default function* rootSaga() {
-    yield all([userSaga()]);
-}
