+++ furniture-client/src/pages/Profile/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
 
-import { AppDispatch } from '../../features/store';
+import { AppDispatch } from '../../store/store';
 
 
 const Profile: React.FC = () => {
