+++ furniture-client/src/store/rootSaga.ts
@@ -1,0 +1,8 @@
+import { all } from 'redux-saga/effects';
+
+import { authUserSaga } from './slice/authUser/authUserSaga';
+
+
+export default function* rootSaga() {
+    yield all([authUserSaga()]);
+}
