+++ furniture-client/src/types/user.interface.ts
@@ -1,0 +1,7 @@
+export interface IUser {
+    id: string;
+    email: string;
+    firstName: string;
+    lastName: string;
+    role: string;
+}
